code formatting

+++ src/org/ddogleg/struct/DogArray_I32.java
@@ -23,7 +23,7 @@ import org.ddogleg.sorting.QuickSort_S32;
 import java.util.Arrays;
 
 /**
- * This is a queue that is composed of integers.  Elements are added and removed from the tail
+ * Growable array composed of ints.
  *
  * @author Peter Abeles
  */
@@ -33,7 +33,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	public int size;
 
 	public DogArray_I32( int maxSize ) {
-		data = new int[ maxSize ];
+		data = new int[maxSize];
 		this.size = 0;
 	}
 
@@ -50,7 +50,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 		return out;
 	}
 
-	public static DogArray_I32 array( int ...values ) {
+	public static DogArray_I32 array( int... values ) {
 		DogArray_I32 out = zeros(values.length);
 		for (int i = 0; i < values.length; i++) {
 			out.data[i] = values[i];
@@ -64,7 +64,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	public int count( int value ) {
 		int total = 0;
 		for (int i = 0; i < size; i++) {
-			if( data[i] == value )
+			if (data[i] == value)
 				total++;
 		}
 		return total;
@@ -77,7 +77,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	 * @return true if equal or false if not
 	 */
 	public boolean isEquals( int... values ) {
-		if (size!=values.length)
+		if (size != values.length)
 			return false;
 		for (int i = 0; i < size; i++) {
 			if (data[i] != values[i])
@@ -92,27 +92,27 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	public void addAll( DogArray_I32 queue ) {
-		if( size+queue.size > data.length ) {
-			int[] temp = new int[ (size+queue.size) * 2];
-			System.arraycopy(data,0,temp,0,size);
+		if (size + queue.size > data.length) {
+			int[] temp = new int[(size + queue.size)*2];
+			System.arraycopy(data, 0, temp, 0, size);
 			data = temp;
 		}
-		System.arraycopy(queue.data,0,data,size,queue.size);
+		System.arraycopy(queue.data, 0, data, size, queue.size);
 		size += queue.size;
 	}
 
-	public void addAll( int[] array , int startIndex , int endIndex ) {
-		if( endIndex > array.length )
-			throw new IllegalAccessError("endIndex is larger than input array. "+endIndex+" > "+array.length);
+	public void addAll( int[] array, int startIndex, int endIndex ) {
+		if (endIndex > array.length)
+			throw new IllegalAccessError("endIndex is larger than input array. " + endIndex + " > " + array.length);
 
-		int arraySize = endIndex-startIndex;
+		int arraySize = endIndex - startIndex;
 
-		if( size+arraySize > data.length ) {
-			int[] temp = new int[ (size+arraySize) * 2];
-			System.arraycopy(data,0,temp,0,size);
+		if (size + arraySize > data.length) {
+			int[] temp = new int[(size + arraySize)*2];
+			System.arraycopy(data, 0, temp, 0, size);
 			data = temp;
 		}
-		System.arraycopy(array,startIndex,data,size,arraySize);
+		System.arraycopy(array, startIndex, data, size, arraySize);
 		size += arraySize;
 	}
 
@@ -121,37 +121,39 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	public void push( int val ) {
-		if( size == data.length ) {
+		if (size == data.length) {
 			int[] temp;
 			try {
-				temp = new int[ size * 2+5];
-			} catch( OutOfMemoryError e ) {
+				temp = new int[size*2 + 5];
+			} catch (OutOfMemoryError e) {
 				System.gc();
-				temp = new int[ 3*size/2];
+				temp = new int[3*size/2];
 			}
-			System.arraycopy(data,0,temp,0,size);
+			System.arraycopy(data, 0, temp, 0, size);
 			data = temp;
 		}
-		data[size++] = val;
+		data[size++] = (int)val;
 	}
 
 	/**
 	 * Sets this array to be equal to the array segment
+	 *
 	 * @param array (Input) source array
 	 * @param offset first index
 	 * @param length number of elements to copy
 	 */
-	public void setTo( int[] array , int offset , int length ) {
+	public void setTo( int[] array, int offset, int length ) {
 		resize(length);
-		System.arraycopy(array,offset,data,0,length);
+		System.arraycopy(array, offset, data, 0, length);
 	}
 
 	/**
 	 * Set's the value of this array to the passed in raw array.
+	 *
 	 * @param src (Input) The input array
 	 * @return A reference to "this" to allow chaining of commands
 	 */
-	public DogArray_I32 setTo( int... src) {
+	public DogArray_I32 setTo( int... src ) {
 		setTo(src, 0, src.length);
 		return this;
 	}
@@ -161,31 +163,32 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	 */
 	public int[] toArray() {
 		int[] out = new int[size];
-		System.arraycopy(data,0,out,0,size);
+		System.arraycopy(data, 0, out, 0, size);
 		return out;
 	}
 
 	public void remove( int index ) {
-		for( int i = index+1; i < size; i++ ) {
-			data[i-1] = data[i];
+		for (int i = index + 1; i < size; i++) {
+			data[i - 1] = data[i];
 		}
 		size--;
 	}
 
 	/**
 	 * Removes elements from the list starting at 'first' and ending at 'last'
+	 *
 	 * @param first First index you wish to remove. Inclusive.
 	 * @param last Last index you wish to remove. Inclusive.
 	 */
-	public void remove( int first , int last ) {
-		if( last < first )
-			throw new IllegalArgumentException("first <= last. first="+first+" last="+last );
-		if( last >= size )
-			throw new IllegalArgumentException("last must be less than the max size. last="+last+" size="+size);
+	public void remove( int first, int last ) {
+		if (last < first)
+			throw new IllegalArgumentException("first <= last. first=" + first + " last=" + last);
+		if (last >= size)
+			throw new IllegalArgumentException("last must be less than the max size. last=" + last + " size=" + size);
 
-		int delta = last-first+1;
-		for( int i = last+1; i < size; i++ ) {
-			data[i-delta] = data[i];
+		int delta = last - first + 1;
+		for (int i = last + 1; i < size; i++) {
+			data[i - delta] = data[i];
 		}
 		size -= delta;
 	}
@@ -193,18 +196,18 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	/**
 	 * Inserts the value at the specified index and shifts all the other values down.
 	 */
-	public void insert( int index , int value ) {
-		if( size == data.length ) {
-			int[] temp = new int[ size * 2+5];
-			System.arraycopy(data,0,temp,0,index);
+	public void insert( int index, int value ) {
+		if (size == data.length) {
+			int[] temp = new int[size*2 + 5];
+			System.arraycopy(data, 0, temp, 0, index);
 			temp[index] = value;
-			System.arraycopy(data,index,temp,index+1,size-index);
+			System.arraycopy(data, index, temp, index + 1, size - index);
 			this.data = temp;
 			size++;
 		} else {
 			size++;
-			for( int i = size-1; i > index; i-- ) {
-				data[i] = data[i-1];
+			for (int i = size - 1; i > index; i--) {
+				data[i] = data[i - 1];
 			}
 			data[index] = value;
 		}
@@ -218,8 +221,8 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	 * @return The removed object
 	 */
 	public int removeSwap( int index ) {
-		if( index < 0 || index >= size )
-			throw new IllegalArgumentException("Out of bounds. index="+index+" max size "+size);
+		if (index < 0 || index >= size)
+			throw new IllegalArgumentException("Out of bounds. index=" + index + " max size " + size);
 		int ret = data[index];
 		size -= 1;
 		data[index] = data[size];
@@ -227,7 +230,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	public int removeTail() {
-		if( size > 0 ) {
+		if (size > 0) {
 			size--;
 			return data[size];
 		} else {
@@ -236,46 +239,47 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	public int get( int index ) {
-		if( index < 0 || index >= size)
-			throw new IndexOutOfBoundsException("index = "+index+"  size = "+size);
+		if (index < 0 || index >= size)
+			throw new IndexOutOfBoundsException("index = " + index + "  size = " + size);
 		return data[index];
 	}
 
 	public int getTail() {
-		if (size==0)
+		if (size == 0)
 			throw new IndexOutOfBoundsException("Array is empty");
-		return data[size-1];
+		return data[size - 1];
 	}
 
 	/**
 	 * Returns an element starting from the end of the list. 0 = size -1
 	 */
 	public int getTail( int index ) {
-		if( index < 0 || index >= size)
-			throw new IndexOutOfBoundsException("index = "+index+"  size = "+size);
-		return data[size-index-1];
+		if (index < 0 || index >= size)
+			throw new IndexOutOfBoundsException("index = " + index + "  size = " + size);
+		return data[size - index - 1];
 	}
 
 	public void setTail( int index, int value ) {
-		if( index < 0 || index >= size)
-			throw new IndexOutOfBoundsException("index = "+index+"  size = "+size);
-		data[size-index-1] = value;
+		if (index < 0 || index >= size)
+			throw new IndexOutOfBoundsException("index = " + index + "  size = " + size);
+		data[size - index - 1] = value;
 	}
 
 	/**
 	 * Gets the value at the index which corresponds to the specified fraction
+	 *
 	 * @param fraction 0 to 1 inclusive
 	 * @return value at fraction
 	 */
 	public int getFraction( double fraction ) {
-		return get( (int)((size-1)*fraction) );
+		return get((int)((size - 1)*fraction));
 	}
 
 	public int unsafe_get( int index ) {
 		return data[index];
 	}
 
-	public void set( int index , int value ) {
+	public void set( int index, int value ) {
 		data[index] = value;
 	}
 
@@ -285,7 +289,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	@Override public void resize( int size ) {
-		if( data.length < size ) {
+		if (data.length < size) {
 			data = new int[size];
 		}
 		this.size = size;
@@ -293,10 +297,11 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 
 	/**
 	 * Resizes the array and assigns the default value to every element.
+	 *
 	 * @param size New size
 	 * @param value Default value
 	 */
-	public void resize( int size , int value ) {
+	public void resize( int size, int value ) {
 		resize(size);
 		fill(value);
 	}
@@ -311,7 +316,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 
 	public boolean contains( int value ) {
 		for (int i = 0; i < size; i++) {
-			if( data[i] == value )
+			if (data[i] == value)
 				return true;
 		}
 		return false;
@@ -325,8 +330,8 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	@Override public void reserve( int amount ) {
 		if (data.length >= amount)
 			return;
-		int []tmp = new int[amount];
-		System.arraycopy(data,0,tmp,0,this.size);
+		int[] tmp = new int[amount];
+		System.arraycopy(data, 0, tmp, 0, this.size);
 		data = tmp;
 	}
 
@@ -335,7 +340,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	@Override public void zero() {
-		Arrays.fill(data,0,size,(int)0);
+		Arrays.fill(data, 0, size, (int)0);
 	}
 
 	@Override public DogArray_I32 copy() {
@@ -345,11 +350,11 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	@Override public void flip() {
-		if( size <= 1 )
+		if (size <= 1)
 			return;
 
 		int D = size/2;
-		for (int i = 0,j=size-1; i < D; i++,j--) {
+		for (int i = 0, j = size - 1; i < D; i++, j--) {
 			int tmp = data[i];
 			data[i] = data[j];
 			data[j] = tmp;
@@ -357,48 +362,49 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	public int pop() {
-        return data[--size];
-    }
+		return data[--size];
+	}
 
 	/**
 	 * Returns the index of the first element with the specified 'value'.  return -1 if it wasn't found
+	 *
 	 * @param value Value to search for
 	 * @return index or -1 if it's not in the list
 	 */
 	public int indexOf( int value ) {
 		for (int i = 0; i < size; i++) {
-			if( data[i] == value )
+			if (data[i] == value)
 				return i;
 		}
 		return -1;
 	}
 
 	public int indexOfGreatest() {
-		if( size <=- 0 )
+		if (size <= -0)
 			return -1;
 
 		int selected = 0;
 		int best = data[0];
 
 		for (int i = 1; i < size; i++) {
-			if( data[i] > best ) {
+			if (data[i] > best) {
 				best = data[i];
 				selected = i;
 			}
 		}
 
-		return selected	;
+		return selected;
 	}
 
 	public int indexOfLeast() {
-		if( size <=- 0 )
+		if (size <= -0)
 			return -1;
 
 		int selected = 0;
 		int best = data[0];
 
 		for (int i = 1; i < size; i++) {
-			if( data[i] < best ) {
+			if (data[i] < best) {
 				best = data[i];
 				selected = i;
 			}
@@ -408,7 +414,7 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 	}
 
 	@Override public void sort() {
-		Arrays.sort(data,0,size);
+		Arrays.sort(data, 0, size);
 	}
 
 	/**
@@ -418,19 +424,19 @@ public class DogArray_I32 implements DogArrayPrimitive<DogArray_I32> {
 		sorter.sort(data,size);
 	}
 
-	public void forIdx(FunctionEachIdx func) {
+	public void forIdx( FunctionEachIdx func ) {
 		for (int i = 0; i < size; i++) {
-			func.process(i,data[i]);
+			func.process(i, data[i]);
 		}
 	}
 
-	public void forEach(FunctionEach func) {
+	public void forEach( FunctionEach func ) {
 		for (int i = 0; i < size; i++) {
 			func.process(data[i]);
 		}
 	}
 
-	public void applyIdx(FunctionApplyIdx func) {
+	public void applyIdx( FunctionApplyIdx func ) {
 		for (int i = 0; i < size; i++) {
 			data[i] = func.process(i, data[i]);
 		}
