+++ src/org/ddogleg/struct/DogArray_I64.java
@@ -21,7 +21,7 @@ package org.ddogleg.struct;
 import java.util.Arrays;
 
 /**
- * This is a queue that is composed of integers.  Elements are added and removed from the tail
+ * Growable array composed of longs.
  *
  * @author Peter Abeles
  */
@@ -31,7 +31,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	public int size;
 
 	public DogArray_I64( int maxSize ) {
-		data = new long[ maxSize ];
+		data = new long[maxSize];
 		this.size = 0;
 	}
 
@@ -48,7 +48,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 		return out;
 	}
 
-	public static DogArray_I64 array( long ...values ) {
+	public static DogArray_I64 array( long... values ) {
 		DogArray_I64 out = zeros(values.length);
 		for (int i = 0; i < values.length; i++) {
 			out.data[i] = values[i];
@@ -62,7 +62,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	public int count( long value ) {
 		int total = 0;
 		for (int i = 0; i < size; i++) {
-			if( data[i] == value )
+			if (data[i] == value)
 				total++;
 		}
 		return total;
@@ -75,7 +75,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	 * @return true if equal or false if not
 	 */
 	public boolean isEquals( long... values ) {
-		if (size!=values.length)
+		if (size != values.length)
 			return false;
 		for (int i = 0; i < size; i++) {
 			if (data[i] != values[i])
@@ -90,27 +90,27 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	public void addAll( DogArray_I64 queue ) {
-		if( size+queue.size > data.length ) {
-			long[] temp = new long[ (size+queue.size) * 2];
-			System.arraycopy(data,0,temp,0,size);
+		if (size + queue.size > data.length) {
+			long[] temp = new long[(size + queue.size)*2];
+			System.arraycopy(data, 0, temp, 0, size);
 			data = temp;
 		}
-		System.arraycopy(queue.data,0,data,size,queue.size);
+		System.arraycopy(queue.data, 0, data, size, queue.size);
 		size += queue.size;
 	}
 
-	public void addAll( long[] array , int startIndex , int endIndex ) {
-		if( endIndex > array.length )
-			throw new IllegalAccessError("endIndex is larger than input array. "+endIndex+" > "+array.length);
+	public void addAll( long[] array, int startIndex, int endIndex ) {
+		if (endIndex > array.length)
+			throw new IllegalAccessError("endIndex is larger than input array. " + endIndex + " > " + array.length);
 
-		int arraySize = endIndex-startIndex;
+		int arraySize = endIndex - startIndex;
 
-		if( size+arraySize > data.length ) {
-			long[] temp = new long[ (size+arraySize) * 2];
-			System.arraycopy(data,0,temp,0,size);
+		if (size + arraySize > data.length) {
+			long[] temp = new long[(size + arraySize)*2];
+			System.arraycopy(data, 0, temp, 0, size);
 			data = temp;
 		}
-		System.arraycopy(array,startIndex,data,size,arraySize);
+		System.arraycopy(array, startIndex, data, size, arraySize);
 		size += arraySize;
 	}
 
@@ -119,37 +119,39 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	public void push( long val ) {
-		if( size == data.length ) {
+		if (size == data.length) {
 			long[] temp;
 			try {
-				temp = new long[ size * 2+5];
-			} catch( OutOfMemoryError e ) {
+				temp = new long[size*2 + 5];
+			} catch (OutOfMemoryError e) {
 				System.gc();
-				temp = new long[ 3*size/2];
+				temp = new long[3*size/2];
 			}
-			System.arraycopy(data,0,temp,0,size);
+			System.arraycopy(data, 0, temp, 0, size);
 			data = temp;
 		}
-		data[size++] = val;
+		data[size++] = (long)val;
 	}
 
 	/**
 	 * Sets this array to be equal to the array segment
+	 *
 	 * @param array (Input) source array
 	 * @param offset first index
 	 * @param length number of elements to copy
 	 */
-	public void setTo( long[] array , int offset , int length ) {
+	public void setTo( long[] array, int offset, int length ) {
 		resize(length);
-		System.arraycopy(array,offset,data,0,length);
+		System.arraycopy(array, offset, data, 0, length);
 	}
 
 	/**
 	 * Set's the value of this array to the passed in raw array.
+	 *
 	 * @param src (Input) The input array
 	 * @return A reference to "this" to allow chaining of commands
 	 */
-	public DogArray_I64 setTo( long... src) {
+	public DogArray_I64 setTo( long... src ) {
 		setTo(src, 0, src.length);
 		return this;
 	}
@@ -159,31 +161,32 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	 */
 	public long[] toArray() {
 		long[] out = new long[size];
-		System.arraycopy(data,0,out,0,size);
+		System.arraycopy(data, 0, out, 0, size);
 		return out;
 	}
 
 	public void remove( int index ) {
-		for( int i = index+1; i < size; i++ ) {
-			data[i-1] = data[i];
+		for (int i = index + 1; i < size; i++) {
+			data[i - 1] = data[i];
 		}
 		size--;
 	}
 
 	/**
 	 * Removes elements from the list starting at 'first' and ending at 'last'
+	 *
 	 * @param first First index you wish to remove. Inclusive.
 	 * @param last Last index you wish to remove. Inclusive.
 	 */
-	public void remove( int first , int last ) {
-		if( last < first )
-			throw new IllegalArgumentException("first <= last. first="+first+" last="+last );
-		if( last >= size )
-			throw new IllegalArgumentException("last must be less than the max size. last="+last+" size="+size);
+	public void remove( int first, int last ) {
+		if (last < first)
+			throw new IllegalArgumentException("first <= last. first=" + first + " last=" + last);
+		if (last >= size)
+			throw new IllegalArgumentException("last must be less than the max size. last=" + last + " size=" + size);
 
-		int delta = last-first+1;
-		for( int i = last+1; i < size; i++ ) {
-			data[i-delta] = data[i];
+		int delta = last - first + 1;
+		for (int i = last + 1; i < size; i++) {
+			data[i - delta] = data[i];
 		}
 		size -= delta;
 	}
@@ -191,18 +194,18 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	/**
 	 * Inserts the value at the specified index and shifts all the other values down.
 	 */
-	public void insert( int index , long value ) {
-		if( size == data.length ) {
-			long[] temp = new long[ size * 2+5];
-			System.arraycopy(data,0,temp,0,index);
+	public void insert( int index, long value ) {
+		if (size == data.length) {
+			long[] temp = new long[size*2 + 5];
+			System.arraycopy(data, 0, temp, 0, index);
 			temp[index] = value;
-			System.arraycopy(data,index,temp,index+1,size-index);
+			System.arraycopy(data, index, temp, index + 1, size - index);
 			this.data = temp;
 			size++;
 		} else {
 			size++;
-			for( int i = size-1; i > index; i-- ) {
-				data[i] = data[i-1];
+			for (int i = size - 1; i > index; i--) {
+				data[i] = data[i - 1];
 			}
 			data[index] = value;
 		}
@@ -216,8 +219,8 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	 * @return The removed object
 	 */
 	public long removeSwap( int index ) {
-		if( index < 0 || index >= size )
-			throw new IllegalArgumentException("Out of bounds. index="+index+" max size "+size);
+		if (index < 0 || index >= size)
+			throw new IllegalArgumentException("Out of bounds. index=" + index + " max size " + size);
 		long ret = data[index];
 		size -= 1;
 		data[index] = data[size];
@@ -225,7 +228,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	public long removeTail() {
-		if( size > 0 ) {
+		if (size > 0) {
 			size--;
 			return data[size];
 		} else {
@@ -234,46 +237,47 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	public long get( int index ) {
-		if( index < 0 || index >= size)
-			throw new IndexOutOfBoundsException("index = "+index+"  size = "+size);
+		if (index < 0 || index >= size)
+			throw new IndexOutOfBoundsException("index = " + index + "  size = " + size);
 		return data[index];
 	}
 
 	public long getTail() {
-		if (size==0)
+		if (size == 0)
 			throw new IndexOutOfBoundsException("Array is empty");
-		return data[size-1];
+		return data[size - 1];
 	}
 
 	/**
 	 * Returns an element starting from the end of the list. 0 = size -1
 	 */
 	public long getTail( int index ) {
-		if( index < 0 || index >= size)
-			throw new IndexOutOfBoundsException("index = "+index+"  size = "+size);
-		return data[size-index-1];
+		if (index < 0 || index >= size)
+			throw new IndexOutOfBoundsException("index = " + index + "  size = " + size);
+		return data[size - index - 1];
 	}
 
 	public void setTail( int index, long value ) {
-		if( index < 0 || index >= size)
-			throw new IndexOutOfBoundsException("index = "+index+"  size = "+size);
-		data[size-index-1] = value;
+		if (index < 0 || index >= size)
+			throw new IndexOutOfBoundsException("index = " + index + "  size = " + size);
+		data[size - index - 1] = value;
 	}
 
 	/**
 	 * Gets the value at the index which corresponds to the specified fraction
+	 *
 	 * @param fraction 0 to 1 inclusive
 	 * @return value at fraction
 	 */
 	public long getFraction( double fraction ) {
-		return get( (int)((size-1)*fraction) );
+		return get((int)((size - 1)*fraction));
 	}
 
 	public long unsafe_get( int index ) {
 		return data[index];
 	}
 
-	public void set( int index , long value ) {
+	public void set( int index, long value ) {
 		data[index] = value;
 	}
 
@@ -283,7 +287,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	@Override public void resize( int size ) {
-		if( data.length < size ) {
+		if (data.length < size) {
 			data = new long[size];
 		}
 		this.size = size;
@@ -291,10 +295,11 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 
 	/**
 	 * Resizes the array and assigns the default value to every element.
+	 *
 	 * @param size New size
 	 * @param value Default value
 	 */
-	public void resize( int size , long value ) {
+	public void resize( int size, long value ) {
 		resize(size);
 		fill(value);
 	}
@@ -309,7 +314,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 
 	public boolean contains( long value ) {
 		for (int i = 0; i < size; i++) {
-			if( data[i] == value )
+			if (data[i] == value)
 				return true;
 		}
 		return false;
@@ -323,8 +328,8 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	@Override public void reserve( int amount ) {
 		if (data.length >= amount)
 			return;
-		long []tmp = new long[amount];
-		System.arraycopy(data,0,tmp,0,this.size);
+		long[] tmp = new long[amount];
+		System.arraycopy(data, 0, tmp, 0, this.size);
 		data = tmp;
 	}
 
@@ -333,7 +338,7 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	@Override public void zero() {
-		Arrays.fill(data,0,size,(long)0);
+		Arrays.fill(data, 0, size, (long)0);
 	}
 
 	@Override public DogArray_I64 copy() {
@@ -343,11 +348,11 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	@Override public void flip() {
-		if( size <= 1 )
+		if (size <= 1)
 			return;
 
 		int D = size/2;
-		for (int i = 0,j=size-1; i < D; i++,j--) {
+		for (int i = 0, j = size - 1; i < D; i++, j--) {
 			long tmp = data[i];
 			data[i] = data[j];
 			data[j] = tmp;
@@ -355,48 +360,49 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	public long pop() {
-        return data[--size];
-    }
+		return data[--size];
+	}
 
 	/**
 	 * Returns the index of the first element with the specified 'value'.  return -1 if it wasn't found
+	 *
 	 * @param value Value to search for
 	 * @return index or -1 if it's not in the list
 	 */
 	public int indexOf( long value ) {
 		for (int i = 0; i < size; i++) {
-			if( data[i] == value )
+			if (data[i] == value)
 				return i;
 		}
 		return -1;
 	}
 
 	public int indexOfGreatest() {
-		if( size <=- 0 )
+		if (size <= -0)
 			return -1;
 
 		int selected = 0;
 		long best = data[0];
 
 		for (int i = 1; i < size; i++) {
-			if( data[i] > best ) {
+			if (data[i] > best) {
 				best = data[i];
 				selected = i;
 			}
 		}
 
-		return selected	;
+		return selected;
 	}
 
 	public int indexOfLeast() {
-		if( size <=- 0 )
+		if (size <= -0)
 			return -1;
 
 		int selected = 0;
 		long best = data[0];
 
 		for (int i = 1; i < size; i++) {
-			if( data[i] < best ) {
+			if (data[i] < best) {
 				best = data[i];
 				selected = i;
 			}
@@ -406,22 +412,22 @@ public class DogArray_I64 implements DogArrayPrimitive<DogArray_I64> {
 	}
 
 	@Override public void sort() {
-		Arrays.sort(data,0,size);
+		Arrays.sort(data, 0, size);
 	}
 
-	public void forIdx(FunctionEachIdx func) {
+	public void forIdx( FunctionEachIdx func ) {
 		for (int i = 0; i < size; i++) {
-			func.process(i,data[i]);
+			func.process(i, data[i]);
 		}
 	}
 
-	public void forEach(FunctionEach func) {
+	public void forEach( FunctionEach func ) {
 		for (int i = 0; i < size; i++) {
 			func.process(data[i]);
 		}
 	}
 
-	public void applyIdx(FunctionApplyIdx func) {
+	public void applyIdx( FunctionApplyIdx func ) {
 		for (int i = 0; i < size; i++) {
 			data[i] = func.process(i, data[i]);
 		}
